Replace `tsc` with `vite` for build
