+++ package.json
@@ -40,7 +40,7 @@
 		},
 		"#*": {
 			"development": "./src/*.ts",
-			"default": "./dist/*.js"
+			"default": "./src/*.ts"
 		}
 	},
 	"exports": {
@@ -51,15 +51,15 @@
 	"scripts": {
 		"build": "pnpm run \"/^build:.*/\" ",
 		"build:doc": "typedoc",
-		"build:pkg": "tsc",
+		"build:pkg": "vite build",
 		"clean": "pnpm run \"/^clean:.*/\"; del \"./**/node_modules\" ",
-		"clean:build": "del \".dist\" ",
+		"clean:build": "del \"./dist\" ",
 		"clean:cache": "del \"./node_modules/.cache\" ",
 		"clean:doc": "del \"./docs\" ",
 		"clean:test": "del \"./coverage\" ",
 		"dev": "pnpm run \"/^dev:.*/\" ",
 		"dev:doc": "typedoc --watch",
-		"dev:pkg": "tsc --watch",
+		"dev:pkg": "vite build --watch",
 		"dev:test": "pnpm vitest watch --passWithNoTests --ui",
 		"fix": "pnpm run \"/^fix:.*/\" ",
 		"fix:format": "biome format . --verbose --write",
@@ -88,6 +88,7 @@
 		"svelte": "5.0.0-next.166",
 		"type-fest": "4.20.1",
 		"typescript": "5.5.2",
+		"vite": "5.3.2",
 		"zimmerframe": "1.1.2"
 	},
 	"peerDependencies": {
