+++ tsconfig.json
@@ -1,11 +1,8 @@
 {
 	"$schema": "https://json.schemastore.org/tsconfig",
-	"extends": ["@total-typescript/tsconfig/tsc/dom/library"],
+	"extends": ["@total-typescript/tsconfig/bundler/dom/library"],
 	"compilerOptions": {
 		"customConditions": ["development"],
-		"rootDir": ".",
-		"outDir": "dist",
 		"types": ["vitest/importMeta"]
-	},
-	"include": ["src/", "test/"]
+	}
 }
